requirements.txt
python-telegram-bot>=20.0
python-dotenv
langgraph
langchain
langchain-google-genai
